+++ libraries/FastShiftIn/FastShiftIn.cpp
@@ -1,7 +1,7 @@
 //
 //    FILE: FastShiftIn.cpp
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.3.0
+// VERSION: 0.3.2
 // PURPOSE: Fast ShiftIn for 74HC165 register, AVR optimized
 //    DATE: 2013-09-29
 //     URL: https://github.com/RobTillaart/FastShiftIn
@@ -14,8 +14,8 @@
 
 FastShiftIn::FastShiftIn(uint8_t dataIn, uint8_t clockPin, uint8_t bitOrder)
 {
-  _bitOrder = bitOrder;
-  _value    = 0;
+  _bitOrder  = bitOrder;
+  _lastValue = 0;
   pinMode(dataIn, INPUT);
   pinMode(clockPin, OUTPUT);
   //  https://www.arduino.cc/reference/en/language/functions/advanced-io/shiftin/
@@ -41,8 +41,7 @@ FastShiftIn::FastShiftIn(uint8_t dataIn, uint8_t clockPin, uint8_t bitOrder)
 }
 
 
-
-int FastShiftIn::read()
+uint16_t FastShiftIn::read()
 {
   if (_bitOrder == LSBFIRST)
   {
@@ -52,8 +51,65 @@ int FastShiftIn::read()
 }
 
 
+uint16_t FastShiftIn::read16()
+{
+  uint16_t rv;
+  if (_bitOrder == LSBFIRST)
+  {
+    rv = readLSBFIRST();
+    rv += uint16_t(readLSBFIRST()) << 8;
+    return rv;
+  }
+  rv = readMSBFIRST();
+  rv <<= 8;
+  rv += readMSBFIRST();
+  return rv;
+}
 
-int FastShiftIn::readLSBFIRST()
+
+uint32_t FastShiftIn::read24()
+{
+  uint32_t rv;
+  if (_bitOrder == LSBFIRST)
+  {
+    rv = readLSBFIRST();
+    rv += uint32_t(readLSBFIRST()) << 8;
+    rv += uint32_t(readLSBFIRST()) << 16;
+    return rv;
+  }
+  rv = readMSBFIRST();
+  rv <<= 8;
+  rv += readMSBFIRST();
+  rv <<= 8;
+  rv += readMSBFIRST();
+  return rv;
+}
+
+
+uint32_t FastShiftIn::read32()
+{
+  uint32_t rv;
+  if (_bitOrder == LSBFIRST)
+  {
+    rv = readLSBFIRST();
+    rv += uint32_t(readLSBFIRST()) << 8;
+    rv += uint32_t(readLSBFIRST()) << 16;
+    rv += uint32_t(readLSBFIRST()) << 24;
+    return rv;
+  }
+  rv = readMSBFIRST();
+  rv <<= 8;
+  rv += readMSBFIRST();
+  rv <<= 8;
+  rv += readMSBFIRST();
+  rv <<= 8;
+  rv += readMSBFIRST();
+  return rv;
+}
+
+
+
+uint8_t FastShiftIn::readLSBFIRST()
 {
 #if defined(ARDUINO_ARCH_AVR) || defined(ARDUINO_ARCH_MEGAAVR)
 
@@ -74,20 +130,20 @@ int FastShiftIn::readLSBFIRST()
     *_clockRegister &= cbmask2;
     SREG = oldSREG;
   }
-  _value = rv;
+  _lastValue = rv;
   return rv;
 
 #else
 
-  // reference implementation
-  _value = shiftIn(_dataPinIn, _clockPin, LSBFIRST);
-  return _value;
+  //  reference implementation
+  _lastValue = shiftIn(_dataPinIn, _clockPin, LSBFIRST);
+  return _lastValue;
 
 #endif
 }
 
 
-int FastShiftIn::readMSBFIRST()
+uint8_t FastShiftIn::readMSBFIRST()
 {
 #if defined(ARDUINO_ARCH_AVR) || defined(ARDUINO_ARCH_MEGAAVR)
 
@@ -108,23 +164,23 @@ int FastShiftIn::readMSBFIRST()
     *_clockRegister &= cbmask2;
     SREG = oldSREG;
   }
-  _value = rv;
+  _lastValue = rv;
   return rv;
 
 #else
 
-  // reference implementation
-  _value = shiftIn(_dataPinIn, _clockPin, MSBFIRST);
-  return _value;
+  //  reference implementation
+  _lastValue = shiftIn(_dataPinIn, _clockPin, MSBFIRST);
+  return _lastValue;
 
 #endif
 
 }
 
 
-int FastShiftIn::lastRead(void)
+uint32_t FastShiftIn::lastRead(void)
 {
-  return _value;
+  return _lastValue;
 };
 
 
@@ -145,5 +201,5 @@ uint8_t FastShiftIn::getBitOrder(void)
 };
 
 
-// -- END OF FILE --
+//  -- END OF FILE --